feat: refactor gene selection table
