--- src/components/PlotSelectionTable/index.ts
@@ -1,2 +1,0 @@
-export { PlotSelectionTable } from './PlotSelectionTable';
-export type { PlotSelectionTableProps, PlotSelectionState } from './PlotSelectionTable.types';
