+++ src/components/ROIGenesSelectionTable/index.ts
@@ -1,0 +1,2 @@
+export { PlotSelectionTable } from './ROIGenesSelectionTable';
+export type { PlotSelectionTableProps, PlotSelectionState } from './ROIGenesSelectionTable.types';
